fixing extra whitespace
